Zygote: Additional whitelisting for legacy devices.
On M and below, we provide a blanket whitelist for all files under
"/vendor/zygote_whitelist". This path is whitelisted purely to allow
this patch to be applied easily on legacy devices and configurations.

Note that this does not amount to a loosening of our security policy
because whitelisted files are reopened anyway.

Bug: 32691930
Test: manual

(cherry picked from commit bbdddde4f90db828bf95899414e4bc3c66225428)

Change-Id: I9700fc7b469d0bc4d876c52292f25888b94a5223