Fully removed the bottom stack
Moved the ambient notifications into the shelf
instead of having them behave with their own
physics. Previously the ambient notifications
were stuck in the old world behind the shelf,
but this doesn't make sense anymore.

This also fixed a bug where the notifications were
wrongly positioned on the lockscreen since the shelf
counted as a view that wasn't gone.

Test: add low-priority notifications, observe scrolling
Bug: 32437839
Change-Id: I3921ea9f80a06f1b6330315423b012174269ac8e