DO NOT MERGE Rework Device Monitoring Dialog in Quicksettings
If a device owner is active, the layout is changed from a
standard AlertDialog with only one string to a custom dialog
that includes information on Device Owners, and VPN and Network Logging,
if enabled.

Cherry-picked from master and modified to not rely on changes that are not in
N:

A change in master renamed some variables in QSFooter.createDialog() and made
them final, this change is also included here.

The owner's organization name cannot be set in N.

In the case of active Device Owner and active VPN, there was a specialized
string for this in N. This string was now replaced by using the simpler
Device Owner string + an own entry for VPN in the new layout.

Some imports had to be included that were already there on master.

The dimen-tag throws an error message on floats during build, replaced by an
item-tag.

BUG: 29748723
BUG: 33126622
Test: Manual, CTS-Verifier tests will be added later

(cherry picked from commit ada85f3695a4d02113608be4c32a06efab503ae8)

Change-Id: I2bfca9d9d02a42d9c3b17683625eda29e9369666