Prepare to enable multiprocess by default.
Implement logic to track the version on which the user last manually
enabled/disabled multiprocess (by storing the version code in the
setting, negated for disabled). This enables us to have the following
logic:

1) If multiprocess is enabled by default, re-enable it for users who
   turned it off on a previous version, but leave it off for users who
   turned it off on the current version. This lets us collect fresh
   bug reports after fixing previous issues.

2) If multiprocess is disabled by default, leave it on for users who
   have enabled it manually, but turn it back off for users who were
   just relying on the default.

This CL leaves it set to "disabled by default" so there is no immediate
behavioural change, but this enables us to switch the "enabled by
default" flag back and forth freely later.

Bug: 21643067
Test: run WebViewUpdateServiceTest
Test: manually upgrade/downgrade webview, verify behaviour of setting
Change-Id: Ibbecac7e5ee75e9c4a1b6efb84253ace98d1eb1a