Add additional logging for temperature readings.
In order to determine how often devices are reaching various
temperatures, we now cache the last hour of temperature readings.
Once it's been an hour, we calculate the avg/min/max and log the results.

Additionally, this CL adds some logging to show how the temperature has
changed over time during the one hour window and what the charging status
is when the notification shows.

Bug: 34510881, 34379897
Test: manual
Change-Id: I62a5ccc92122f5d2955a52677740e896f0d7892b