Revert "Fix window transformation related issue"
This reverts commit 7d745078e9a93b5b620ecf2f3d997ca10379354e.

This commit is being reverted because the logic it tries to fix does not
make sense and caused a regression causing all animation positions to be
offset incorrectly but applying the animation metrics before applying the
window offset (mainly for bottom screen in multiwindow). The original
case this tries to fix is for older apps like gallery2 which crashes
(ramdump) when trying reproduce the steps.

Fixes: 34769878
Test: ran original case that this tried to fix and played with recents
related to the new bug it created